create outcomes helper
